requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pandas
numpy
faiss-cpu
sentence-transformers
scikit-learn
streamlit
python-dotenv
google-generativeai